# public/swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/seed/admin": {
      "post": {
        "operationId": "AuthGeneralController_seedAdmin",
        "summary": "Seed first account",
        "description": "Create an admin account with provided data if correct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create an admin account with provided data if correct",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Already created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth",
          "v1/seed"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/account/guest": {
      "post": {
        "operationId": "AuthGeneralController_createGuestAccount",
        "summary": "Create an guest account with correct access key",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuestAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create an guest account with correct access key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Misssing key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth",
          "v1/guest"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/user": {
      "get": {
        "operationId": "AuthGeneralController_getUser",
        "summary": "Get Logged in user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get Logged in user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorize",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/optional-user": {
      "get": {
        "operationId": "AuthGeneralController_getOptionalUser",
        "summary": "Get Logged in user optional",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get Logged in user optional",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Not available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/check-email": {
      "post": {
        "operationId": "AuthGeneralController_isExistingUser",
        "summary": "Check email exist in database",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Is existing user",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "existinguser": true
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/update": {
      "post": {
        "operationId": "AuthGeneralController_updateUser",
        "summary": "Update an account with provided data if correct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create an account with provided data if correct"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/register": {
      "post": {
        "operationId": "AuthGeneralController_register",
        "summary": "Create an account with provided data if correct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create an account with provided data if correct",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/login": {
      "post": {
        "operationId": "AuthGeneralController_login",
        "summary": "Logs in user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logs in user",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "token": "token"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth",
          "v1/login"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/refresh-token": {
      "post": {
        "operationId": "AuthGeneralController_refreshToken",
        "summary": "Refersh token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "token": "token"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh user token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "accessToken": "token"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/forgot-password": {
      "post": {
        "operationId": "AuthGeneralController_forgetPassword",
        "summary": "Sent forget-password otp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sent forget-password otp",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "message": "An email is sent at XXXXXXXXXXXXXX"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/receive-otp": {
      "post": {
        "operationId": "AuthGeneralController_receiveOtp",
        "summary": "Verify otp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verify otp",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "isVerified": true,
                    "message": "Otp verification success"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid otp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/otp-newpassword": {
      "post": {
        "operationId": "AuthGeneralController_setNewPassword",
        "summary": "Set new password with otp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordOtpAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Set new password with otp",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "message": "User password change success"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid otp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/isloggedin": {
      "get": {
        "operationId": "AuthGeneralController_isLoggedin",
        "summary": "Check User Login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Check User Login",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/logout": {
      "get": {
        "operationId": "AuthGeneralController_logout",
        "summary": "Logs out user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logs out user",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/local/role": {
      "get": {
        "operationId": "AuthGeneralController_getRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get user role",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "role": [
                      "user"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "v1/auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Boilerplate v1 API",
    "description": "Basic docs for API in version 1",
    "version": "1.0",
    "contact": {
      "name": "Shirsho ",
      "url": "https://marketron.com",
      "email": "hniqbal01@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ExceptionResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Invalid request parameters"
          },
          "timestamp": {
            "type": "string",
            "example": "2024-11-22T04:10:35.729Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "CreateAuthDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Shirsho"
          },
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "rolePermission": {
            "type": "string",
            "example": "user"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User name",
            "example": "shirsho"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "shirsho@gmail.com"
          },
          "phone": {
            "type": "string",
            "description": "User phone",
            "example": "1234567890"
          },
          "imageUrl": {
            "type": "string",
            "description": "User image url",
            "example": "image.jpg"
          },
          "description": {
            "type": "string",
            "description": "User description",
            "example": "I am a developer"
          },
          "role": {
            "description": "User role",
            "example": [
              "USER"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rolePermission": {
            "type": "string",
            "description": "User role permission",
            "example": "user"
          },
          "isverified": {
            "type": "boolean",
            "description": "User is verified",
            "example": false
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "imageUrl",
          "description",
          "role",
          "rolePermission",
          "isverified"
        ]
      },
      "CreateGuestAccountDto": {
        "type": "object",
        "properties": {
          "accountType": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "key": {
            "type": "string",
            "example": "11ece8be27491e39b6f940c18fa339ce4d2c9dadd94064898035e16ec27e7f76"
          }
        },
        "required": [
          "accountType",
          "key"
        ]
      },
      "EmailAuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Shirsho"
          },
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "phone": {
            "type": "string",
            "example": "01534115844"
          },
          "status": {
            "type": "string",
            "example": "active"
          },
          "imageUrl": {
            "type": "string",
            "example": "imageUrl"
          },
          "description": {
            "type": "string",
            "example": "description"
          },
          "provider": {
            "type": "string",
            "example": "Google"
          },
          "providerId": {
            "type": "string",
            "example": "local"
          },
          "role": {
            "example": "user",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rolePermission": {
            "type": "string",
            "example": "user"
          }
        }
      },
      "LoginAuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "VerifyEmailAuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          },
          "removeotp": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "email",
          "otp",
          "removeotp"
        ]
      },
      "ChangePasswordOtpAuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "otp",
          "password"
        ]
      },
...
    }
  }
}